# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [orders]
      summary: Hold a table
      description: >
        Creates a pending order that holds the table for the payment window.
        The deposit is priced on the server from the zone's minimum spend and
        the party size — the client never sends an amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Table held, awaiting payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /ticket-orders:
    post:
      operationId: createTicketOrder
      tags: [orders]
      summary: Hold event tickets
      description: >
        Creates a pending order that holds the tickets for the payment window.
        The total is priced on the server from the event's ticket price.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketOrderRequest"
      responses:
        "201":
          description: Tickets held, awaiting payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Orders for a phone number
      description: Backs the "My tickets" screen. Newest first.
      parameters:
        - name: phone
          in: query
          required: true
          description: Thai mobile number; punctuation is ignored.
          schema:
            type: string
            minLength: 9
            maxLength: 20
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderKind"
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/{code}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: One order by its reference
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{code}/slip:
    post:
      operationId: submitSlip
      tags: [orders]
      summary: Submit a transfer slip
      description: >
        Records the SHA-256 of the uploaded slip and marks the order paid. A
        hash that has already been spent is rejected with 409 — the uniqueness
        is enforced by the database, not by the caller.
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSlipRequest"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlipResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          description: The payment window closed and the hold was released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /orders/{code}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      description: >
        Releases the held table or tickets. The booking phone number must match
        — the reference alone is not treated as proof of ownership.
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: The phone number does not match the order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  responses:
...
    BadRequest:
      description: The request was rejected
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: The request conflicts with the current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      description: Every non-2xx response in this API has this shape.
      properties:
        error:
          type: string
          description: Stable machine-readable code, e.g. `table_taken`.
        message:
          type: string
          description: Human-readable Thai text, safe to show to a guest.
      required:
        - error
        - message
    Slot:
      type: string
      description: Seating slot, HH:mm.
      enum:
        - "18:00"
        - "19:00"
        - "20:00"
        - "21:00"
        - "22:00"
        - "23:00"
...
    OrderKind:
      type: string
      enum: [table, ticket]
    OrderStatus:
      type: string
      enum: [pending, paid, cancelled, expired]
...
    Order:
      type: object
      description: >
        One shape for both kinds. `guests`/`tableId`/`zoneId`/`zoneName` are
        present only for table bookings; `eventId`/`eventTitle`/`quantity` only
        for ticket orders.
      properties:
        code:
          type: string
        kind:
          $ref: "#/components/schemas/OrderKind"
        status:
          $ref: "#/components/schemas/OrderStatus"
        bookerName:
          type: string
        phone:
          type: string
        email:
          type: string
        occasion:
          type: string
        notes:
          type: string
        date:
          type: string
        slot:
          type: string
        holdUntil:
          type: string
        guests:
          type: integer
        tableId:
          type: string
        zoneId:
          type: string
        zoneName:
          type: string
        eventId:
          type: string
        eventTitle:
          type: string
        quantity:
          type: integer
        amount:
          type: number
          description: What the guest pays, VAT included, in THB.
        amountBase:
          type: number
          description: VAT-exclusive portion of `amount`.
        amountVat:
          type: number
        holdExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
      required:
        - code
        - kind
        - status
        - bookerName
        - phone
        - date
        - slot
        - holdUntil
        - amount
        - amountBase
        - amountVat
        - holdExpiresAt
        - createdAt
    CreateBookingRequest:
      type: object
      properties:
        bookerName:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 9
          maxLength: 20
        email:
          type: string
        guests:
          type: integer
          minimum: 1
          maximum: 20
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        slot:
          $ref: "#/components/schemas/Slot"
        tableId:
          type: string
        holdUntil:
          type: string
          pattern: "^\\d{2}:\\d{2}$"
        occasion:
          type: string
          maxLength: 60
        notes:
          type: string
          maxLength: 500
      required:
        - bookerName
        - phone
        - guests
        - date
        - tableId
    CreateTicketOrderRequest:
      type: object
      properties:
        bookerName:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 9
          maxLength: 20
        email:
          type: string
        eventId:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 10
        notes:
          type: string
          maxLength: 500
      required:
        - bookerName
        - phone
        - eventId
        - quantity
    SubmitSlipRequest:
      type: object
      properties:
        hash:
          type: string
          description: Lowercase hex SHA-256 of the slip image.
          pattern: "^[0-9a-f]{64}$"
        fileName:
          type: string
          maxLength: 255
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 5242880
      required:
        - hash
        - fileName
        - sizeBytes
    SlipResult:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        slipStatus:
          type: string
          enum: [pending_bank_check, verified, rejected]
        pendingBankCheck:
          type: boolean
          description: >
            True until a slip-verification provider is wired up. The duplicate
            check is real; authenticity is not yet checked.
      required:
        - order
        - slipStatus
        - pendingBankCheck
    CancelOrderRequest:
      type: object
      properties:
        phone:
          type: string
          minLength: 9
          maxLength: 20
      required:
        - phone
...